refactor!: fully migrate to web api using srvx
