+++ package.json
@@ -42,6 +42,7 @@
   "dependencies": {
     "cookie-es": "^2.0.0",
     "rou3": "^0.5.1",
+    "srvx": "^0.2.1",
     "uncrypto": "^0.1.3"
   },
   "resolutions": {
