+++ src/event.ts
@@ -1,8 +1,21 @@
-import type { H3Event, H3EventContext } from "../../types";
-import type { H3EventResponse } from "../../types/event";
-import { BaseEvent } from "../base/event";
+import type { H3Event, H3EventContext, HTTPMethod } from "./types";
+import type { H3EventResponse } from "./types/event";
 
-export class WebEvent extends BaseEvent implements H3Event {
+const H3EventContext = /* @__PURE__ */ (() => {
+  const C = function () {};
+  C.prototype = Object.create(null);
+  return C;
+})() as unknown as { new (): H3EventContext };
+
+const HeadersObject = /* @__PURE__ */ (() => {
+  const C = function () {};
+  C.prototype = Object.create(null);
+  return C;
+})() as unknown as { new (): H3EventContext };
+
+export class H3WebEvent implements H3Event {
+  static __is_event__ = true;
+
   request: Request;
   response: H3EventResponse;
 
@@ -13,9 +26,19 @@ export class WebEvent extends BaseEvent implements H3Event {
   _queryString?: string;
 
   constructor(request: Request, context?: H3EventContext) {
-    super(context);
+    this.context = context || new H3EventContext();
     this.request = request;
     this.response = new WebEventResponse();
+  }
+
+  context: H3EventContext;
+
+  get method(): HTTPMethod {
+    return this.request.method as HTTPMethod;
+  }
+
+  get headers(): Headers {
+    return this.request.headers;
   }
 
   get url() {
@@ -25,7 +48,7 @@ export class WebEvent extends BaseEvent implements H3Event {
     return this._url;
   }
 
-  override get path() {
+  get path() {
     return this.pathname + this.queryString;
   }
 
@@ -76,13 +99,15 @@ export class WebEvent extends BaseEvent implements H3Event {
     }
     return this._queryString;
   }
-}
 
-const HeadersObject = /* @__PURE__ */ (() => {
-  const C = function () {};
-  C.prototype = Object.create(null);
-  return C;
-})() as unknown as { new (): H3EventContext };
+  toString(): string {
+    return `[${this.request.method}] ${this.request.url}`;
+  }
+
+  toJSON(): string {
+    return this.toString();
+  }
+}
 
 class WebEventResponse implements H3EventResponse {
   _headersInit?: Record<string, string>;
